Small improvements

+++ components/stream_server/stream_server.h
@@ -39,7 +39,7 @@ public:
     int get_client_count() { return this->clients_.size(); }
 
     void setRegisterUint16(uint8_t unit, uint8_t function, uint16_t address, uint16_t value, uint16_t maxage);
-    void setRegisterSint32(uint8_t unit, uint8_t function, uint16_t address, float value, uint16_t maxage);
+    void setRegisterSint32(uint8_t unit, uint8_t function, uint16_t address, int32_t value, uint16_t maxage);
 
 protected:
     void accept();
